S220-013 Allow Terminating annotation as package aspect
The Terminating annotation can now be specified in aspect form in addition
to the pragma form, on both subprograms and packages. It already worked on
subprograms, but required some changed to work on packages.

* docs/ug/en/source/how_to_use_gnatprove_in_a_team.rst
* docs/ug/en/source/how_to_write_subprogram_contracts.rst
Document the aspect form of annotations for GNATprove, for both Terminating
and for justifying check messages.

* gnat2why/spark/spark_definition.adb
(Mark_Pragma_Annot_In_Pkg): Take into account annotations that follow the
 package spec.

* testsuite/gnatprove/tests/S220-013__terminating/test.out
Update test output.

Change-Id: I1108f364fcc452a9850f44237061e22c45948776