Q612-030 workaround incorrect generation of component label in gnatwhy3
This hides a crash caused by the interaction of optimization of
new eval_match which removes record with one field, 'First/'Last
projections, removal of us__split_fields which all somehow makes it
impossible to have a correct order of the names given to 'Last/'First
elements. Example: 2646.2579.2354'First should be 2646.2354.2579'First
to be correctly typed by the new gnat2why collection of counterexamples.
This commit solves the problem by replacing an assert with an if (and
subsequently allowing ill formed counterexample in parts of the code). In
this case, counterexamples will disappear for this line.

* gnat2why/why/gnat2why-counter_examples.adb
(Build_Variables_Info): Transform the assert into an if and when the ill
formed case happen, skip the counterexample.
(Add_Attributes): Safety guard on the call to Is_Visible_In_Type

-* O224-038__inlining
No change to this test but the crash involved appear in that file which
now shows no counterexample for the concerned line.

Change-Id: Ic1e2cf41af7c91a23c9d835de2bb5a8cc7dc56c0