use context when searching type of a let binding
no-tn-check

In the Why3 AST inside gnat2why, let bindings have a field for their Why
type. The constructor has Why_Empty as a default value for this field.
When searching for the type of a let binding, we used to look at the type
stored in the node for the let binding. This resulted in a crash when the
type was not supplied during the construction. To alleviate this problem,
we now look at the type of the context field in the let binding if the
type field of the let binding is Why_Empty.

* gnat2why/why/why-gen-expr.adb
(Get_Type): On a W_Binding_Id, first look at the Typ component. If it is
 Why_Empty, call Get_Type recursively on its context.

Change-Id: I4a8afc8d1a1f0ce3e2ff4004c4e22ef60b4a01ae