Q612-030 newsystem update normal and large testsuite
This contains all that was not updated before. In particular, missing
checks and XFAIL mode will be added by this commit.

* M326-011__trace
Problem of traces -> XFAIL

* N120-055__perm
The .mlw was switched to newsystem. Some checks seem to disappear: I am not
sure what they correspond to in the past version. This should be
investigated as part of newsystem ticket.

* N122-021__succ_floats
Check at 30:19 is provable by an older version of altergo (1.30). This
removes the necessity to prove the last check. This is an ok regression.

* N616-037__vscomp2014
A check disappear. The split of gnat now accidentally removes one
attribute containing the name of the check. The reason is likely to be a
new optimization because all the goal generated are concluded by "True"
(investigated with manual proof).
Acceptable regression.

* N715-015__flow_oo_basic
The removed check cannot actually be proved entirely (it seems) because it
appears in the output 3 lines later. This is better than before.

* O129-025__unsound_axiom
The postcondition is clearly unreachable from the code. It is normal that
it is optimized away by Why3. Note that the postcondition is
"Post => False".

* O227-007__float_double_inconsistency
Works locally with 20000 steps: increasing steps to 30000 for nightly
build.

* P219-012__proof_state
It seems that output-msg-only is broken. Adding XFAIL with explanation.

* P530-022__loopinv
Large test to update -> XFAIL because it gets to the time limit. A solution
should be found in the newsystem ticket.

* P715-019__flow_subprogram_termination
This ouput is probably a merge error. This is part of flow analysis which
is indeed correct. The call is done after "OS_Exit(0);" and has return
type No_Return.

* QA11-003__itp_sumrange
sumrange is broken. XFAIL ->To be solved in the ticket for newsystem

* QB01-015__counterexample
The counterexample is not generated because a let binding is not
simplified. To overcome this problem we can tag *temp* let binding with
[@mlw:proxy_symbol] so that Why3 knows it can delete them. XFAIL: To be
solved in newsystem ticket.

* R125-020__comparison
Minor diff in counterexamples. Updating output.

* R416-004__frame_condition
There was a duplication of range check proved. These corresponds to the
following statement lvalue: A (31 .. 40) (I .. I). It is likely that some
part of the former goals is now simplified. The check reappear if the
second I is (I+1) so I think this is optimized away thus ok.
Minor counterexamples diff.

* R509-011__higher_order*
Failing but was able to update them locally. To be done with a different
setup.

* RA29-047__flow_array_crash
Number of provers per goal change.

* RC17-022__private_pointer
Provability change: one change of prover.

* S117-042__flow_case
The two checks that disappear are clearly never reached. They are most
likely optimized away by the tool. Accepting the diff.

* S219-014__nested_update
Minor counterexamples diff and prover diff in goals.

* adacore_u
2 more regressions. Some counterexamples and provers diff. This is updated
from the mailserver: no more diff to be expected.
Accepting the diffs.

* allocators
XFAIL -> to reupdate sessions and output

* random
Last check only provable with a local version of altergo. To be updated.

* red_black_trees
Too large diffs. XFAIL: Reupdate the session.

* sparkskein
One check fails. XFAIL: Reupdate the session when possible

* string_search
XFAIL: reupdate the session

* text_io_get_line
One more regression in this test. Acceptable.

* tictactoe
XFAIL: reupdate the session/output

Change-Id: I70bc8285631d0f380aba905c2cb9c354086f018e