S304-027 support array element conversion in left values
Fix the return type of the sliding function for conversions between
array types with different component types. Also declare array conversion
functions for the return conversion, in case the conversion is used for
a left value.

* gnat2why/why/gnat2why-expr.adb
(Transform_Expr): When an array conversion is encountered, generate
 conversion modules for both directions at once.

* gnat2why/why/why-gen-expr.adb
(Insert_Array_Conversion): Return type of the slide function is the split
 form of the from type, not the to type.

* tests/S304-027__array_conv
File test.

Change-Id: I94722cd41a17c10b047ab0e1702bd3e8b8f61426