S306-041 Update test for realizations
The errors were caused by why3realize not being able to find the Why3
theories. This adds the corresponding -L and also add some checking that
why3realize (in particular) is successful. Previously, when why3realize
failed, it was very easy to miss it because the generated Coq file were
still present.
Also, reverse the order of compilation of stdlib between Map and Set.

* O702-009__coq_realizations/test.out
Add a new more precise output

* O702-009__coq_realizations/test.py
Add printing statement, add why3_lib as -L, and print failing statement
when the subprocess exit code is not 0

Change-Id: Id37bcfb5c24392a05f5c3b7d4544ba6706ec1115