Updated git submodules
Project: why3 master 322376d21cac85aece1e9855c1ba5e65fcf33d09

S306-026 Restore the behavior of output-msg-only

output-msg-only works by rerunning gnatwhy3 without running any actual
proof. So, it uses the result of non-obsolete goals to give a detailed
summary. It uses the checksums of goals to give this summary so it only
works when it has just been run locally (checksums are inside shapes
files).
This commit removes a change I added to get shapes that would not differ
too much in Why3 (for our testsuite): when checksum are equal and shapes
differ (this can happen because explanation are inside shapes but not in
the checksum or because the version of shapes changes), it still
makes the goal obsolete if the shapes are not the same.
In this example, the shapes differ with the same checksum, so the proof is
made obsolete (but tasks are the same). Actually, the precondition and
postconditions are exactly the same task except for the explanation.

This commit removes the line making those obsolete.
This would probably be a better idea to disable this via a debug flag...
or to change the way pairing is done (but I think we had a good reason for
not changing the pairing when this change was done).

* src/session/termcode.ml
Don't make the goal obsolete when shapes are different but checksums are
the same.

Change-Id: I9f3f58c7393cdc99494d995f37b8b6706ba3cf1c