R522-061, S308-011 mark subtype declarations in top-down style
For a subtype declaration like:

   subtype S is T;

we now call Mark_Type_Entity on S and examine the SPARK status of T from
within that routine, not the other way round. This ensures that violations
are correctly propagated from T to S. AFAIR, this was difficult with linear
marking, but seems entirely intuitive with the current top-down scheme.

* gnat2why/spark/spark_definition.adb
(Mark): do not mark ancestor subtype after the In_SPARK status of the newly
 created type has been decided anyway.

(Mark_Type_Entity): instead of blindly calling Mark_Entity just to pull the
 ancestor to translation we now check if the ancestor is in SPARK and
 propagate violation if it isn't.

* testsuite/gnatprove/tests/R329-026__subtype/test.out
Violation is now reported on the subtype, not on the ancestor; this seems
acceptable.

* testsuite/gnatprove/tests/R522-061__weird_symbol/test.out
Weird symbol in marking is gone; violation is reported in a better place,
i.e. in unit P which has SPARK_Mode aspect and not in unit Pulled, which
has no SPARK_Mode at all. Note: this patch avoids the use of Original_Node,
which was discussed as a plausible solution, but was not convincing enough.

Change-Id: I41fc9c175519786684ac4fcea014ce23083c9803