S308-011 do not mark types outside of Mark_Type_Entity
When seeing a type declaration in Mark all we should do was is to call
Mark_Entity, because this is where the SPARK status is decided. Actually, it
only required a semantically-neutral refactor in Transform_Declaration.

* gnat2why/spark/spark_definition.adb
(Mark): only call Mark_Entity when seeing a type declaration.

* gnat2why/why/gnat2why-expr.adb
(Transform_Declaration): move reassignment of Base to IF-statement body,
 because it is only referenced from that body (otherwise we would crash in
 on some assertion related to Retysp).

Change-Id: I4800e12fc2f271266cb9534ac89c9a7959eb2d1f