S320-040 add test about trig function being potentially blocking
This is a minimal reproducer based on the test that I attempted on S310-005,
but here the trigonometric function is called indirectly, i.e. via ProxyX,
which is essential to trigger to spurious checks.

Note that on Proxy2 we always get a check, because it has a local instance
of a complex generic package and thus is never inlined. On Proxy1 we only
get a check when inlining is explicitly disabled.

* testsuite/gnatprove/tests/S320-040__flow_blocking/
Add a minimal reproducer; checks about Sin being a potentially blocking
subprogram are spurious.

Change-Id: Ib951a85cd34dff21ddb2d1fac88f6106ba9dce15