S314-023 detect out of memory exit of gnatwhy3
Ocaml has two ways of exiting in case of out of memory:
1) raising an exception
2) "fatal error: out of memory" is printed on stderr by the runtime,
   and the program stops

In case (1), gnatwhy3 will return with a JSON record that indicates
the error situation, and we already deal with that. The code added
here deals with (2). Before attempting to JSON-parse the string
returned by gnatwhy3, we check if it's the string "fatal error: out
of memory", and report an out of memory situation in that case.

* gnat2why-driver.adb
(Run_Gnatwhy3): capture stderr as well as stdout
* gnat2why-error_messages.adb
(Parse_Why3_Results): detect specific error string and stop in that
  case

Change-Id: I6401958ac85ab6d1e87a80241da830a83bfc27d2