S306-029 Update session for higher order proof
This commits tries to update and make it easier to understand and update
higher_order_proof in the future. In particular, it tries to split the
package instances in several files (to be able to do Prove File quickly as
Prove Line/Check do not work because of nested generics).

* testsuite/gnatprove/R509-011__higher_order_proof/test-higher_order*.ads
Split the files in 4 different files to help future update for this test.
I did this because it is impossible to do Prove Check/Prove Line (nested
generics). It helps and it now takes less time to do Prove File on one of
the tests.

* testsuite/gnatprove/R509-011__higher_order_proof/test.py
Add new files to the stuff to be proved. Add describing comments from the
.ads files.
Add a way to print "FAIL" when there is not exactly 9 medium conditions
(corresponding to the 9 axioms).

* testsuite/gnatprove/R509-011__higher_order_proof/*/*.xml
Update session

Change-Id: I31e62f0ee984d78313648d75fdfa511e1d480c8a