S326-020 fix flattening of objects with pointer types
When assigning a record aggregate to a dereferenced pointer, e.g.

   P.all := (A => ..., B => ..., C => ...);

the RHS is rightly represented as a map of components=>values. It seems
reasonable to also represent the LHS as a map and this patch does this, as
it goes in direction of flow-analysing pointers as precisely as possible.

* gnat2why/flow/flow_utility.adb
(Flatten_Variable): access types are now transparent.

* testsuite/gnatprove/tests/QA23-042__flow_pointers/
Enable primary test.

* testsuite/gnatprove/tests/RB14-019__subtype_checks/test.out
Update output; component of a pointed-to object is now recognized just like
components of ordinary objects.

Change-Id: I38fae3c8e09db4c70f4031bfbb31c8bd94480412