S326-003 consistent handling of pragma Unreferenced and junk names
Warning about ineffective statements was already suppressed for statements
that assign objects annotated as Unreferenced. Now this warning is also
suppressed when assigning to objects with "junk names". This makes GNATprove
(the prover) more consistent with GNAT (the compiler).

* [private] testsuite/gnatprove/tests/S326-003__flow_dummy_loop/
Remove pramga Unreferenced and rename a dummy loop parameter to "Unused";
the warning about statement having no effect is gone.

* testsuite/gnatprove/tests/NC18-019__flow_nested_package_bodies/test.out
Update output. The code in question is:

   package P
     with Initializes => Used
   is
      Used   : Integer := 0;  --  OK
      Unused : Integer := 0;  --  ineffective initialization
   end Nested;

but note that we still complain about Unused not being mentioned in the
Initializes contract.

Change-Id: Iea69e28831fe8a25e3b4c3493b96a48cb795f0cf