Updated git submodules
Project: why3 master 4df80c1501cf304f1e20981bdb5f6f4e0b258b44

QC06-011 Adapt b18ffc3 for Why3 master

This adapts the ConnectionError exception to output a nice message in Why3

* src/driver/prove_client.ml
Register the new exception so that it is nicely printed.

Change-Id: I9483fd0bf2561eaed007432b1eb982e7fd086486


QC06-011 improve error when socket file cannot be created

When the socket file cannot be created, raise an exception instead of
exiting.  This allows a nice error message to be printed.

* prove_client.ml
(client_connect): in case of error, do not exit, but raise exception

Change-Id: I83fb36a431bf4d854a23d93b0b870fa3d0db7526