S322-026 simplify detection of base protected types
If parent of a type entity is an N_[Protected|Task]_Type_Declaration then
the defined entity must be of E_[Protected|Task]_Type kind. In terms of
logic, the condition was "A and B" and we had "B implies A"; this is now
simplified to "B".

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): simplify condition for base concurrent types.

Change-Id: I9195c5e343d9cc06f9d6c47eb5617baba900af5b