RC11-011 Cleanup library directories when using switch --clean
For a library project, ALI files might be copied to either one of the
following directories specified in the project file:
- Library_Dir for the external interface of the library
- Library_ALI_Dir specifically for ALI files

GNATprove will generate sub-directories "gnatprove" in each, which need
to be deleted when GNATprove is called with the switch --clean.
Now fixed

* gnatprove/configuration.adb
(Clean_Up): Remove generated "gnatprove" sub-directory in both library
 directories, for each project file.

* testsuite/gnatprove/tests/RC11-011__standalone_library/incr.adb
* testsuite/gnatprove/tests/RC11-011__standalone_library/test.gpr
* testsuite/gnatprove/tests/RC11-011__standalone_library/test.out
* testsuite/gnatprove/tests/RC11-011__standalone_library/test.py
Add test.

Change-Id: Ia49949c267081c8dbb5e9362e7ff1192ac75f6bd