Updated git submodules
Project: why3 master 98f2c606c908a402516eecc913832ca69d19cfdb

S414-004 Fix realizations

This regenerate realization (probably a diff with Why3 in the Coq
printer).

* lib/coq/int/Power.v
Add a newline at end of file

Change-Id: I3200034c928d87ae2bf5b17adaccc1bc5be45a6e


S412-008 Fix memoization of checksum

This solves a problem with the memoization of the function that performs
checksum of tasks.

* src/session/termcode.ml
(task_v2): Remove argument [version] which prevents the closure.
(task_v3): The complete function has to be duplicated with CV3.

Change-Id: I6d9b8f44ee6f339a24c243565cb74d8997d0d080