S325-002 symbols of access to the full view of an incomplete type
Access types designating the same subtype share a common representative
type (because of anonymous access types). If an access type designates a
type which is the full view of an incomplete type, its symbols will be
share with the access type designating the incomplete view. When looking
for these symbols, search in the appropriate module depending on whether
or not the representative pointer type designate an incomplete type.

* gnat2why/why/why-atree-modules.adb
(Insert_Type_Symbols): Look for symbols in the appropriate module depending
 on whether or not the representative pointer type designates an incomplete
 type.

* S325-002__anonymous_access
New test.

Change-Id: I305235322f085006edb1202bc19b7a72d3a92d81