RC17-019 fix test N804-036__bad_prover
Follow up on R326-064 in which we now always print a warning when a prover
is executed and raise HighFailure.
This commit fixes the configuration of N804-036__bad_prover so that it
uses transformation that exists and that gnatwhy3 only fails when it
launches the prover (raising the warning in test.out) not before during
execution of the driver.

Note that we could detect errors in the application of the driver too but
it was not the purpose of this ticket.

* testsuite/gnatprove/tests/N804-036__bad_ptover/plop.drv
Copy the current altergo driver

* testsuite/gnatprove/tests/N804-036__bad_prover/test.out
The prover error now appear


Change-Id: I8daf32f1b8be4e3c51b472a3cbd38cc3e6026ba2