S423-033 Move pointer ownership checking to gnat2why
While the code of the pointer ownership checking remains in GNAT
frontend, the actual call now resides in gnat2why, so that we can
build on the SPARK_Mode boundary knowledge in gnat2why.

This patch was created by Claire Dross.

* gnat2why/spark/spark_definition.adb
(Get_Emit_Messages): Function reading the value of Emit_Messages.
(Safe_Retysp): Mark the entity before calling Retysp on it. It is used
 inside ownership checking because it does not follow the same structure
 as regular marking and therefore may encounter unmarked nodes.
(Ownership_Checking): Instance of Sem_SPARK.
(Mark_Compilation_Unit): Check rules for pointer aliasing on the whole
 compilation unit.
(Mark_Function_Specification): Perform checks for traversal functions.
(Mark_Type_Entity): Perform checks for ownership types. Do not reject
 access types declared in a part with mode Auto anymore.

* testsuite/gnatprove/tests/P831-045__pointers
Fix test so that it compiles and goes to marking.

* testsuite/gnatprove/tests/RC17-022__private_pointer
Remove an erroneous ASSERT:FAIL marker.

* testsuite/gnatprove/tests/S117-008__pointer
Fix violation of an ownership rule which was not detected before.

* tests/O416-002__dynamic_predicate_not_in_spark/test.out
* tests/RA12-025__pointers/test.out
* tests/RA12-026__split_pointers/test.out
* tests/RA12-032__pointer_free/test.out
* tests/RA29-010__pointer/test.out
* tests/RB05-016__pointer/test.out
* tests/S116-005__pointer/test.out
* tests/S123-026__pointer_aggregate/test.out
* tests/S325-002__rec_pointers_bad/test.out
* tests/TU__derived_types_and_classes_illegal/test.out
Upate output.

Change-Id: Ic8437aa710028e5858f778737c4cff450736b3b6