S423-007 remove an always-true conjunct with Original_Node
Routine Original_Node behaves either as an identity function (for nodes that
have not been rewritten) or returns a non-empty node (for nodes that have
been rewritten). Note: if the intention was to check for rewritten nodes,
then it can be done with Is_Rewrite_Substitution.

* gnat2why/spark/spark_definition.adb
(Mark): remove an always-true conjunct.

Change-Id: Ie8975b7250c0e9a62cface14fbbfa21d87af6c82