S423-033 update expected outputs in disabled test for pointers
One file in this test is meant to be rejected in phase 1, while others in
phase 2. This must be taken into account in the test.py/test.gpr files.

Otherwise, accept current outputs on non-crashing files.

* testsuite/gnatprove/tests/QA23-042__error_pointers/test.out
This test crashes on spark02.adb, but the outputs on other units should be
legitimate, so commit them.

Change-Id: Ib671d97ca51ce67014105fc1f164d1d34522b329