minor do not lock/unlock string table for the gnat2why backend
Apparently gnat2why doesn't change the string table, so there is no need to
lock/unlock it. This locking/unlocking was added in one of the very first
commits, as part of a dummy backend that printed the source-view of the AST,
which probably did some string manipulation.

(no-tn-check)

* gnat2why/why/back_end.adb
(Call_Back_End): do not lock/unlock string table.

Change-Id: I386b8ff07121de01b5b4b01423bbce5c1407000c