S514-011 forget about current incomplete type when resetting spark_mode
In Mark_Entity, we reset the SPARK_Mode to the SPARk_Mode of the entity.
When doing so, we should also reset global variables used to force the
SPARK_Mode. We already did it for delayed type aspects, now do the same
for incomplete types.

* gnat2why/spark/spark_definition.adb
(Mark_Entity): Reset Current_Incomplete_Type when setting current SPARK
 pragma.

* tests/S514-011__pointers/test.out
Test is accepted now.

* tests/S325-002__rec_pointers_bad/test.out
Do not report messages on auto part anymore.

Change-Id: Id1ef3551731c34ad70ce4b249f500a54f3c6f8c1