remove dead code from Mark_Subtype_Indication
This is a follow-up after a similar commit which removed a dead IF-condition
from Mark_Subtype_Indication, but I failed to notice a similarly dead code
in the same routine.

(no-tn-check)

* gnat2why/spark/spark_definition.adb
(Mark_Subtype_Indication): this routine is only called from Mark and always
gets an N_Subtype_Indication parameter (as demonstrated by the added
precondition), so the IF-expression was always returning the THEN branch.

Change-Id: I9ff983b1c51c41465833dfbcb7f9c7499e3c3ea0