more precise locations for type ranges with variable inputs
For an illegal type declaration like:

   X : Integer := 100; -- variable
   type Bounded_Positive is Integer range 1 .. X; -- variable input in range

the error was emitted at "range". Now it is emitted at "X".

Also, this is a prerequisite for adding a precondition to Get_Variables that
will only allow this routine to be called on subexpressions.

(no-tn-check)

* gnat2why/flow/flow-analysis-sanity.adb
(Check_Subtype_Constraints): separately check low and high bound expressions
 of a range constraints.

* testsuite/gnatprove/tests/K609-030__nonstatic_bound/test.out
* testsuite/gnatprove/tests/MB27-015__constraints_var_free/test.out
* testsuite/gnatprove/tests/NA18-003__subp_var_input/test.out
* testsuite/gnatprove/tests/O804-004__inlined_renamed_actuals_problem/...
* testsuite/gnatprove/tests/P517-019__flow_types/test.out
Update locations of flow error messages.

Change-Id: I8954258ff316c024c911c5b6cfbd33c71dc19008