S503-015 Recursive printing of Why3 theories
Instead of printing all theories, we print only the theories that are
recursively included starting from the theories in the "main" section.

As theories are defined by module IDs, but include declarations only
contain symbols, we need a mapping from symbols to theory
declarations.

External axioms (EA) break this new way of printing. If we hit a EA
during the recursive traversal, we fall back to the previous way of
printing. To know if a modules is supposed to be defined by EA, we
need another map.

* gnat2why-driver.adb
(Print_Why_File): get a plan before printing, if the plan is empty
  fall back to previous way of printing
* gnat2why-external_axioms.adb
(Is_External_Axiom_Module): new function to determine if a module is
  defined by an external axiomatization.
(Register_External_Entities): register modules that should be
  covered by the external axiomatization as coming from EA.
* gnat2why-util.adb
(Build_Printing_Plan): new function to compute list of modules to
  print. Returns empty list if the list cannot be computed due to
  EA.
* why-atree-sprint.adb
(Print_Modules_List): new function to print a list of modules
* why-inter.adb
(Find_Decl): new function to find the theory declaration that
  corresponds to a module ID

Change-Id: I73ea09046f067d29533da692f08519f5045a5232