S416-010 Update output
All trivial check will now be tagged with Trivial instead of Cvc4. This is
because we register the is_trivial transformation in Why3. Note that this
transformation was already used on all goals so this should improve
timings (probably negligible improvement)

* testsuite/gnatprove/tests/*/*.out
Add Trivial as prover

* testsuite/gnatprove/QA11-003__itp_server_test/*
Fix the test for QA11-003__itp_server_test: the number of nodes change
because of trivial nodes.

Change-Id: I7839504c3cca782a9c66ca04b63c3a13a4f5aca1