fix mishandling of calls to protected operations in Get_Variables
This fix makes no difference to the test outputs, because it only affects
the narrowing of slocs which happen in First_Variable_Use. Yet, it both
fixes the intended semantics of Get_Variables and removes one call to
Do_Subprogram_Call, which is where the problematic Fold_Function context
component is used.

(no-tn-check)

* gnat2why/flow/flow_utility.adb
(Get_Variables_Internal): when seeing a selected component as a prefix of
 a call to protected operation we should only consider that prefix and not
 the entire call (with its parameters).

Change-Id: I4c2b447b2a913ba0b3a99e26c967c3000cbab156