again more precise locations for type ranges with variable inputs
For an illegal declaration like:

   X : Integer := 123;                    -- variable
   type T4 is array (1 .. X) of Integer;  -- variable input in range

the error was emitted at ".." and now it is emitted at "X". If the same
variable input appear in both low and high bounds, we will get reports for
each bound separately.

Also, this is a prerequisite for adding a precondition to Get_Variables that
will only allow this routine to be called on subexpressions.

(no-tn-check)

* gnat2why/flow/flow-analysis-sanity.adb
(Check_Subtype_Constraints): separately check low and high bound expressions
of a ranges.

* testsuite/gnatprove/tests/P517-019__flow_types/test.out
Update sloc.

Change-Id: I7e348fbd42842072342d8d112a3d533ea4b0a35f