S529-012 do not pull incomplete types if not necessary
As incomplete types (and partial views) used as designated types of access
types are marked later, they can introduce violations that cannot be
recovered from if we assume that they are in SPARK and they are not. Do
not try pulling these entities if they are declared in the private part of
a package with mode Auto.

* gnat2why/spark/spark_util.ads
(Is_Declared_In_Private): Return True if an entity is declared in the
 private part of a package.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): We reject access types designating incomplete types
 or partial views declared in the private part of a package if they have
 no SPARK pragma.

* testsuite/gnatprove/tests/S325-002__rec_pointers_bad/test.out
We have no error as we do not pull the private view of C_Acc anymore.

Change-Id: I02aa730eb09a3dfb8a6502741d1f0bdd64a20cf0