S529-012 reject access subtypes of private general access types
When we have an access subtype, we go to the base type to see if it is
a general access type. If the base type is a private type, we need to
check its underlying type. Note that we assume here that the frontend
will only introduce an access subtype if the subtype declaration has
visibility over the full view of its base type.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Go to underlying type to know if it is a general
 access type.

* tests/S529-012__general_pointer
File test.

Change-Id: Ia5b603ab54d6785618f8f71ff2953dc0750f7acd