Updated git submodules
Project: why3 master a4e93970cf5b71f3eed9c86c67e123b2f9562c6a

S530-009 Change Exit exception in trivial_true

Replace the exception [Exit] in [trivial_true] by something that is not
[is_fatal] (see session_itp.ml).
This avoids the warning to appear when reloading the session: if pairing
fails, a transformation [trivial_true] could be tried on a goal on which
it should fail. In this part of the code, a warning is raised.
Making the error not fatal should make the warning disappear.

* src/transform/gnat_trivial.ml
(is_trivial): Replace Exit with Arg_trans.

Change-Id: Id8e86f18cfd02ab1521ebbcf19eca520fa69fe35