S531-010 avoid unnecessary updates on split records and accesses
When transforming an assignment on a pointer dereference, it is not
necessary to modify the address and is_null fields. In the same way,
when a component of a split record is assigned, it is not necessary to
update its discriminants even if they are mutable.
The change has no impacts on the testsuite as preserved discriminants are
accounted for in the generated frame condition, and we don't have tests
for frame condition on accesses.

* gnat2why/why/gnat2why-expr.adb
(New_Assignment): When assigning to a component of a split record or to the
 value of a split access, preserve the other parts of the object.

Change-Id: Ib8159def9016d40e2ce81e734173754e6682a6f9