S531-019 add test with subprogram identifier in an object name
Apparently we only had tests where prefixes of object name included package
identifiers, e.g. "Pkg_A.Obj_X". This is a test where the prefix is a
subprogram name, like "Proc_A.Obj_X".

It is a minimized version of ACATS test with the same name.

* testsuite/gnatprove/tests/S531-019__flow_expanded_name/
Add; output is as expected.

Change-Id: Idc190ad6b5b3079c8062d4df85f7adb073d3a1cd