R511-031 Add model_trace on old contract variables
This patch adds model_trace on old variable so that we are able to get
a counterexample for them.

* gnat2why-counterexamples.adb
(Build_Variables_Info): Add a case when the slice is 'Old
(same as 'Loop_entry).
Modify 'Loop_Entry case to get outside the condition of
"Contains(Attribute)" which should be unrelated to it.

* gnat2why/why/gnat2why-expr.adb
(Name_For_Old): Add a model_trace to elements that are both old and
inside a contract case.
Add comments to exclude expressions from the Loop_Entry and Old attribute:
only use entities.

* gnat2why/why/why-gen-names.ads
(Old_Label): Definition for Old label.

* testsuite/gnatprove/*/*/test.out
Update tests

Change-Id: I0c090b8dc9284c5e78bfa67e8151dcd90bd0770e