S531-019 do not traverse into N_Expanded_Name
When processing an N_Expanded_Name it is enough to pick Entity (if present).
We don't need to traverse into Prefix, because all we can find there is a
sequence of N_Identifiers for packages/subprograms.

This allows us to have a sanity-checking assertion in Do_Entity, where a
subprogram entity is only expected when traversing statement and when
traversing expressions we only see such entities on function calls (which
have their own handling).

* gnat2why/flow/flow_utility.adb
(Get_Variables_Internal.Proc): do not descend into syntactic children of
 N_Expanded_Name (i.e. Prefix) and N_Identifiers (i.e. Chars).

* testsuite/gnatprove/tests/S531-019__flow_expanded_name/test.opt
Enable, no longer crashes on assertion in Do_Entity.

Change-Id: I574900f26764dfa32a16f5ac66c1d1f63042f5fc