S528-030 Move Refine_Aux to ce_pretty_printing
Refine_Aux is not a mutually recursive function anymore. It is used on a
simple type (all information is in the arguments). So, it could be removed
from the logic of the main program for collecting of counterexamples.
This also changes the name of Refine_Aux to Print_Cntexmp_Value.
This also moves Dont_Display and Make_Trivial as a side effect.

* gnat2why/counterexamples/ce_pretty_printing.ad*
(Print_Cntexmp_Value): Refine_Aux moved. The comment on this function was
updated too.
(Dont_Display): Moved here
(Make_Trivial): Moved here

* gnat2why/counterexamples/gnat2why-counter_examples.adb
(Dont_Display): moved to ce_pretty_printing
(Make_Trivial): moved to ce_pretty_printing
(Refine_Aux): moved to ce_pretty_printing as Print_Cntexmp_Value.

Change-Id: Ib34b38b785c432ee3d8a440b88877a8bb50d4b07