S528-030 Create a file for the data structure of name ordered variables
In the code for counterexamples, there is a map-like datastructure with
keys of the form (Entity_Id, Unbounded_String) and elements containing a
structured counterexample. The important feature of this structure is the
ordering on names (Unbounded_String).
This commit moves the datastructure to a new file (inspired by container
code). The intention of this commit is to make sure that the code has
exactly the same behavior as before (the idea is that the reader should be
able to check that by inlining name_ordered_entities).
The improvement is the removal of container logic the main counterexamples
thus simplifying it.

* gnat2why/counterexamples/gnat2why-counter_example.adb
(Variables_Info): Moved to name_ordered_entities
(Build_Pretty_Line): Use functions defined in name_ordered_entities. The
only difference in complexity should be the variables Values which is not
a renaming anymore: data is get, edited then set again. We could solve
this by passing a pointer but this would be less natural for few
improvements (it seems).

* gnat2why/counterexamples/name_ordered_entities.*
Only necessary functions (for former code to work) were defined in these
files.
(Info): This is always String_CNT_Elements.Map.
(First,Next,Iterator,Cursor_Access,Variable_Info_Access,Iterate,
 Has_Element): Defined to allow the for-loop on the data structure.
(Variables_Info): Kept the name for clarity of the commit.
(Get_Info): Return the info corresponding to the cursor.
(Name): Return the Name from the cursor
(Id): Return the Entity_Id from the cursor
(Insert_Info): Add the info at key given by Id and Name.
(Set_Info): modify the info located at cursor.

Change-Id: I52cd5c5b73721946dbee94a2e2753586503696dd