S613-026 Fix crash on loops with nested exit statements
When a loop contains multiple exits in sequence, GNATprove generates
an ill-formed Why3 code due to an optimization for increasing the
precision of the computation of loop effects, which leads to a crash.

When multiple loops are nested, with an inner loop being unrolled,
the optimization for exits may also lead to crashes.

* gnat2why/why/gnat2why-expr-loops-exits.adb
(Record_And_Replace): Do not perform the optimisation on exits inside
 blocks that will be hoisted due to another exit statement. Document
 better the current algorithm.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): Only push an environment for exit handling
 when the loop is not unrolled.
(Unroll_Loop): Syntactically equivalent rewrite to use one variable.

* testsuite/gnatprove/tests/S613-026__loop_exit/*
Test passes now.

Change-Id: Ibff948b542d446f9ee7a86d0aed1f3ce2f3468c1