S415-012 support local borrowers
We support local borrowers of deep objects by havocing borrowed objects
at the end of the scope of the borrower. Information about the
modifications done during the borrow is carried to the borrowed object
using a predicate that we call pledge. A pledge describes how the borrowed
object and the borrower are related. It can be modified during the traversal
(in case of reborrows). We encode this predicate in Why as an abstract type
along with a get function taking the borrowed object and the borrower as
parameters and returning a boolean. When we havoc a borrowed object, we
assume that the pledge holds by assuming (get pledge borrowed borrower).

* spark/spark_definition.adb
(Has_Deep_Subcomponents_With_Predicates): Return true on types which
 contain deep types which have a subtype predicate. This is used to disallow
 borrowing such types as it can break the predicate arbitrarily nestedly,
 making it difficult to verify.
(Mark_Object_Entity): Local borrowers are now supported, except when they
 are of a type with subcomponents of a deep type which has a subtype
 predicate.

* spark/spark_util.adb
(Is_Local_Borrower): New function returning true for variables or constants
 of an anonymous access to variable type.

* why/gnat2why-decls.adb
(Translate_Variable): For local borrowers, declare a reference for the
 pledge.

* why/gnat2why-expr.adb
(Shift_Rvalue): Moved outside of New_Assignment for reuse, takes an
 additional domain parameter to work on terms.
(New_Pledge_Update_From_Assigment): Create an assignment updating the
 pledge of a local borrower after a borrow or a reborrow.
(Assignment_Of_Obj_Decl): Update pledge if we are declaring a borrower.
(Havoc_Borrowed_Expression): Generate an havoc statement for the borrowed
 object. It assumes the pledge of the borrower.
(New_Assignment): Shift_Rvalue is factored out.
(Transform_Assignment_Statement): Update pledge in case of reborrow.
(Transform_Block_Statement): Havoc borrowed expressions when we exit the
 scope of a borrower.

* why/gnat2why-subprograms.adb
(Compute_Effects): Add effect on pledge if we have a write effect on a
 borrower.
(Generate_VCs_For_Subprogram): Havoc borrowed expressions when we exit the
 scope of a borrower.

* why/gnat2why-util.adb
(Get_Borrows_From_Decls): Returns all the borrowers from a list of
 declarations.

* why/spark_atree-entities.ads
(Is_Anonymous_Access_Type): New wrapper on einfo utility.

* why/why-atree-modules.adb
(Pledge): New module in ada_model for pledges.

* why/why-gen-arrays.adb
(Get_Entity_Of_Variable): Use actual conversion instead of +.

* why/why-gen-pointers.adb
(Borrow_Infos): New data structure holding information for all borrowers
 encountered so far. We store the the borrowed expression, the type of the
 borrower, and identifiers from the associated pledge module, namely, the
 pledge reference and the get function.
(Declare_Pledge_Ref): Clone the appropriate pledge module and emit a
 global ref declaration for the pledge of a borrower.
(Get_Borrowed_Expr): Query the Borrow_Infos map and return the borrowed
 expression.
(Get_Borrowed_Ty): Query the Borrow_Infos map and return the borrowed type.
(Get_Pledge_Id): Query the Borrow_Infos map and return the identifier of the
 pledge reference.
(New_Ada_Pointer_Update): Do not emit a sequence if not in the prog domain.
(New_Pledge_Call): Construct a call to the pledge of a borrower.
(New_Pledge_Update): Construct an update to the pledge of a borrower.

* share/spark/theories/ada__model.mlw
(Pledge): New module for pledges. Morally, a pledge is a predicate
 representing the link between the borrowed value and the borrower. We do
 not use Why3 lambas to avoid introducing polymorphism. Instead, we
 represent a pledge as a two parameter map with a function get to querry
 whether the relation holds between the borrowed value and the borrower.

* tests/K222-004__package
* tests/K222-004__subprogram
* tests/OC17-029__package_body_statements_marking
* tests/P831-045__pointers
* tests/QA23-042__error_pointers
* tests/RA29-007__mutable_borrowers
* tests/RA29-010__pointer
* tests/RB05-016__pointer
Update test outputs, local borrowers are supported now.

* tests/S415-012__illegal_borrow
* tests/S415-012__list
New tests.

Change-Id: I0e953e0b6ea0833313cc153fb83183b7b0d462e1