S614-030 Detect violations of ownership rules in with'ed units
With'ed units that are not explicitly marked in or out of SPARK may
still contain declarations that violate ownership rules. In that case,
we should consider the corresponding entities as not in SPARK to avoid
errors later during flow analysis or proof. This is done by exposing
a function Is_Legal from the ownership checking code, that only
verifies rules not related to permissions, and does not by itself issue
any error message. This is called from marking on declarations that
might violate ownership rules, currently only declarations of variables
and constants.

* gnat2why/spark/spark_definition.adb
(Mark_Entity): Add detection of ownership legality rules on variable/
 constant declarations in SPARK_Mode Auto.

* testsuite/gnatprove/tests/P208-037__interrupt_priority_in_spark/test.opt
* testsuite/gnatprove/tests/P208-037__interrupt_priority_in_spark/test.out
Test passes now.

Change-Id: I4e47c1eac25669501b82f88cb4eab3f4a552fe67